% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check required columns in a data frame}
\usage{
check_columns(df, required_cols)
}
\arguments{
\item{df}{A data frame.}

\item{required_cols}{Character vector of required column names.}
}
\value{
Invisible TRUE if all columns present; otherwise an error is thrown.
}
\description{
Check required columns in a data frame
}
\examples{
\dontrun{
check_columns(mydata, c("Pid", "Date"))
}
}
