% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_plot.R
\name{cumulative_plot}
\alias{cumulative_plot}
\title{Cumulative Article Count Plot}
\usage{
cumulative_plot(cleaned_data, output = NULL)
}
\arguments{
\item{cleaned_data}{Data frame with a `Date` column.}

\item{output}{Optional file path to save PNG.}
}
\value{
A ggplot object.
}
\description{
Creates a line plot of cumulative number of articles per year.
}
